Div småfikser. Leverer poeng som double. Mulig det fikser noen problemer.

+++ sandvika-activity-board-backend/src/main/java/no/bouvet/sandvika/activityboard/api/ActivityController.java
@@ -122,7 +122,7 @@ public class ActivityController
 
         return new ArrayList<>(entries.values())
                 .stream()
-                .sorted(Comparator.comparingInt(LeaderboardEntry::getPoints).reversed())
+                .sorted(Comparator.comparingDouble(LeaderboardEntry::getPoints).reversed())
                 .collect(Collectors.toList());
     }
